RPi4: Add nic component

+++ components/LogServer/src/custom_log_format.c
@@ -63,7 +63,8 @@ _Log_format_convert(
         OS_Logger_LOG_LEVEL_LENGTH, entry->emitterMetadata.filteringLevel,
         OS_Logger_LOG_LEVEL_LENGTH, entry->consumerMetadata.filteringLevel,
 
-        msg_len, entry->msg);
+        (int)msg_len, entry->msg);
+        // msg_len, entry->msg);
 
     return OS_SUCCESS;
 }
